+++ pytreeclass/_src/code_build.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 """Constructor code generation from type annotations."""
 
 from __future__ import annotations
