+++ pytreeclass/_src/tree_base.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 """Define a class that convert a class to a JAX compatible tree structure"""
 
 from __future__ import annotations
@@ -18,7 +19,7 @@ from __future__ import annotations
 import abc
 from collections.abc import Callable
 from contextlib import contextmanager
-from typing import Any, Generic, Hashable, NamedTuple, TypeVar
+from typing import Any, Hashable, NamedTuple, TypeVar
 
 import jax
 import jax.numpy as jnp
@@ -671,297 +672,3 @@ def treeclass_pp(node: TreeClass, **spec: Unpack[PPSpec]) -> str:
     skip = [f.name for f in fields(node) if not f.repr]
     kvs = ((k, v) for k, v in vars(node).items() if k not in skip)
     return name + "(" + pps(kvs, pp=attr_value_pp, **spec) + ")"
-
-
-def _frozen_error(opname: str, tree):
-    raise NotImplementedError(
-        f"Cannot apply `{opname}` operation to a frozen object `{tree!r}`.\n"
-        "Unfreeze the object first to apply operations to it\n"
-        "Example:\n"
-        ">>> import jax\n"
-        ">>> import pytreeclass as pytc\n"
-        ">>> tree = jax.tree_map(pytc.unfreeze, tree, is_leaf=pytc.is_frozen)"
-    )
-
-
-class _Frozen(Generic[T]):
-    __slots__ = ["__wrapped__", "__weakref__"]
-    __wrapped__: T
-
-    def __init__(self, tree: T) -> None:
-        object.__setattr__(self, "__wrapped__", tree)
-
-    def __setattr__(self, _, __) -> None:
-        raise AttributeError("Cannot assign to frozen instance.")
-
-    def __delattr__(self, _: str) -> None:
-        raise AttributeError("Cannot delete from frozen instance.")
-
-    def __repr__(self) -> str:
-        return "#" + tree_repr(self.__wrapped__)
-
-    def __str__(self) -> str:
-        return "#" + tree_str(self.__wrapped__)
-
-    def __copy__(self) -> _Frozen[T]:
-        return _Frozen(tree_copy(self.__wrapped__))
-
-    def __eq__(self, rhs: Any) -> bool | jax.Array:
-        if not isinstance(rhs, _Frozen):
-            return False
-        return is_tree_equal(self.__wrapped__, rhs.__wrapped__)
-
-    def __hash__(self) -> int:
-        return tree_hash(self.__wrapped__)
-
-    # raise helpful error message when trying to interact with frozen object
-    __add__ = __radd__ = __iadd__ = lambda x, _: _frozen_error("+", x)
-    __sub__ = __rsub__ = __isub__ = lambda x, _: _frozen_error("-", x)
-    __mul__ = __rmul__ = __imul__ = lambda x, _: _frozen_error("*", x)
-    __matmul__ = __rmatmul__ = __imatmul__ = lambda x, _: _frozen_error("@", x)
-    __truediv__ = __rtruediv__ = __itruediv__ = lambda x, _: _frozen_error("/", x)
-    __floordiv__ = __rfloordiv__ = __ifloordiv__ = lambda x, _: _frozen_error("//", x)
-    __mod__ = __rmod__ = __imod__ = lambda x, _: _frozen_error("%", x)
-    __pow__ = __rpow__ = __ipow__ = lambda x, _: _frozen_error("**", x)
-    __lshift__ = __rlshift__ = __ilshift__ = lambda x, _: _frozen_error("<<", x)
-    __rshift__ = __rrshift__ = __irshift__ = lambda x, _: _frozen_error(">>", x)
-    __and__ = __rand__ = __iand__ = lambda x, _: _frozen_error("and", x)
-    __xor__ = __rxor__ = __ixor__ = lambda x, _: _frozen_error("xor", x)
-    __or__ = __ror__ = __ior__ = lambda x, _: _frozen_error("or", x)
-    __neg__ = __pos__ = __abs__ = __invert__ = lambda x: _frozen_error("unary op", x)
-    __call__ = lambda x, *_, **__: _frozen_error("call", x)
-
-
-jtu.register_pytree_node(
-    nodetype=_Frozen,
-    flatten_func=lambda tree: ((), tree),
-    unflatten_func=lambda treedef, _: treedef,
-)
-
-
-def freeze(wrapped: _Frozen[T] | T) -> _Frozen[T]:
-    """Freeze a value to avoid updating it by `jax` transformations.
-
-    Example:
-        >>> import jax
-        >>> import pytreeclass as pytc
-        >>> import jax.tree_util as jtu
-        >>> # Usage with `jax.tree_util.tree_leaves`
-        >>> # no leaves for a wrapped value
-        >>> jtu.tree_leaves(pytc.freeze(2.))
-        []
-
-        >>> # retrieve the frozen wrapper value using `is_leaf=pytc.is_frozen`
-        >>> jtu.tree_leaves(pytc.freeze(2.), is_leaf=pytc.is_frozen)
-        [#2.0]
-
-        >>> # Usage with `jax.tree_util.tree_map`
-        >>> a= [1,2,3]
-        >>> a[1] = pytc.freeze(a[1])
-        >>> jtu.tree_map(lambda x:x+100, a)
-        [101, #2, 103]
-
-    Note:
-        - `freeze` is idempotent, i.e. `freeze(freeze(x)) == freeze(x)`
-        - Use `tree_freeze` for the `tree_map` version of `freeze`.
-            Example:
-                >>> import pytreeclass as pytc
-                >>> tree = [1,2,3]
-                >>> # freezing the pytree items
-                >>> pytc.tree_freeze(tree)
-                [#1, #2, #3]
-                >>> # freeze the pytree itself
-                >>> pytc.freeze(tree)
-                #[1, 2, 3]
-    """
-    return wrapped if is_frozen(wrapped) else _Frozen(wrapped)  # type: ignore
-
-
-def is_frozen(wrapped: Any) -> bool:
-    """Returns True if the value is a frozen wrapper."""
-    return isinstance(wrapped, _Frozen)
-
-
-def unfreeze(wrapped: _Frozen[T] | T) -> T:
-    """Unfreeze `frozen` value, otherwise return the value itself.
-
-    - use `is_leaf=pytc.is_frozen` with `jax.tree_map` to unfreeze a tree.**
-
-    Example:
-        >>> import pytreeclass as pytc
-        >>> import jax
-        >>> frozen_value = pytc.freeze(1)
-        >>> pytc.unfreeze(frozen_value)
-        1
-        >>> # usage with `jax.tree_map`
-        >>> frozen_tree = jax.tree_map(pytc.freeze, {"a": 1, "b": 2})
-        >>> unfrozen_tree = jax.tree_map(pytc.unfreeze, frozen_tree, is_leaf=pytc.is_frozen)
-        >>> unfrozen_tree
-        {'a': 1, 'b': 2}
-    """
-    return getattr(wrapped, "__wrapped__") if is_frozen(wrapped) else wrapped
-
-
-def is_nondiff(wrapped: Any) -> bool:
-    """
-    Returns True if the node is a non-differentiable node, and False for if the
-    node is of type float, complex number, or a numpy array of floats or
-    complex numbers.
-
-    Example:
-        >>> import pytreeclass as pytc
-        >>> import jax.numpy as jnp
-        >>> pytc.is_nondiff(jnp.array(1))  # int array is non-diff type
-        True
-        >>> pytc.is_nondiff(jnp.array(1.))  # float array is diff type
-        False
-        >>> pytc.is_nondiff(1)  # int is non-diff type
-        True
-        >>> pytc.is_nondiff(1.)  # float is diff type
-        False
-
-    Note:
-        This function is meant to be used with `jax.tree_map` to
-        create a mask for non-differentiable nodes in a tree, that can be used
-        to freeze the non-differentiable nodes before passing the tree to a
-        `jax` transformation.
-    """
-    if hasattr(wrapped, "dtype") and np.issubdtype(wrapped.dtype, np.inexact):
-        return False
-    if isinstance(wrapped, (float, complex)):
-        return False
-    return True
-
-
-def tree_freeze(
-    tree: T,
-    mask: T | Callable[[Any], bool] = is_nondiff,
-    *,
-    is_leaf: IsLeafType = None,
-):
-    """Freeze a tree of values to avoid updating it by `jax` transformations.
-
-    Args:
-        tree: A pytree of values.
-        mask: A pytree of boolean values or a callable that accepts a leaf and
-            returns a boolean. If a leaf is True either in the mask or the
-            callable output , the leaf is frozen, otherwise it is unchanged.
-            defaults to `pytc.is_nondiff` which freezes all non-differentiable nodes.
-        is_leaf: A callable that accepts a leaf and returns a boolean. If
-            provided, it is used to determine if a value is a leaf. for example,
-            `is_leaf=lambda x: isinstance(x, list)` will treat lists as leaves
-            and will not recurse into them.
-
-    Example:
-        >>> import pytreeclass as pytc
-        >>> tree = [1, 2, {"a": 3, "b": 4.0}]
-        >>> # freeze all non-differentiable nodes by default
-        >>> # the frozen values are prefixed with `#`
-        >>> pytc.tree_freeze(tree)
-        [#1, #2, {'a': #3, 'b': 4.0}]
-        >>> pytc.tree_freeze(tree, mask=lambda x: x>2)
-        [1, 2, {'a': #3, 'b': #4.0}]
-
-    Note:
-        - freezing a value wraps it with a wrapper that yields no leaves when
-            `jax.tree_util.tree_flatten` is called on it.
-        - `tree_freeze` uses `jax.tree_util.tree_map` that applies freezing to
-            all leaves of the tree. This means that if a leaf is a container
-            (e.g. list, dict, tuple, etc.) the container itself is not frozen,
-            but its leaves are. whereas `pytreeclass.freeze` freezes the container
-            itself.
-        - `tree_freeze` common use case is to freeze non-differentiable
-            in conjunction with `tree_unfreeze` to handle non-differentiable
-            nodes in `jax` transformations.
-
-            Example:
-                >>> # pass non-differentiable nodes to `jax.grad`
-                >>> import pytreeclass as pytc
-                >>> import jax
-                >>> @jax.grad
-                ... def square(tree):
-                ...     tree = pytc.tree_unfreeze(tree)
-                ...     return tree[0]**2
-                >>> tree = (1., 2)  # contains a non-differentiable node
-                >>> square(pytc.tree_freeze(tree))
-                (Array(2., dtype=float32, weak_type=True), #2)
-    """
-
-    if isinstance(mask, Callable):
-        return jax.tree_map(
-            lambda x: freeze(x) if mask(x) else x,
-            tree,
-            is_leaf=is_leaf,
-        )
-
-    if jtu.tree_structure(tree, is_leaf=is_leaf) == jtu.tree_structure(mask):
-        return jax.tree_map(
-            lambda x, y: freeze(x) if y else x,
-            tree,
-            mask,
-            is_leaf=is_leaf,
-        )
-
-    raise ValueError(
-        f"`mask` must be a callable that accepts a leaf a returns a boolean "
-        f"or a tree with the same structure as tree with boolean leaves."
-        f" Got {mask=} and {tree=}."
-    )
-
-
-def tree_unfreeze(tree: T, mask: T | Callable[[Any], bool] = lambda _: True):
-    """Unreeze a tree of values that were frozen by `tree_freeze`/`freeze`.
-
-    Args:
-        tree: A pytree of values.
-        mask: A pytree of boolean values or a callable that accepts a leaf and
-            returns a boolean. If a leaf is True either in the mask or the
-            callable return , the leaf is unfrozen, otherwise it is unchanged.
-            defaults to `lambda _: True` which unfreezes all nodes.
-    Example:
-        >>> import pytreeclass as pytc
-        >>> tree = [1, 2, {"a": 3, "b": 4}]
-        >>> pytc.tree_unfreeze(pytc.tree_freeze(tree))
-        [1, 2, {'a': 3, 'b': 4}]
-
-    Note:
-        - `tree_unfreeze` uses `jax.tree_util.tree_map` that applies unfreezing
-            to all leaves of the tree. This means that if a leaf is a container
-            (e.g. list, dict, tuple, etc.) the container itself is not unfrozen,
-            but its leaves are. whereas `unfreeze` unfreezes the container.
-        - `tree_unfreeze` common use case is to unfreeze non-differentiable
-            nodes that were frozen by `tree_freeze` to be passed to `jax`
-
-            Example:
-                >>> # pass non-differentiable nodes to `jax.grad`
-                >>> import pytreeclass as pytc
-                >>> import jax
-                >>> @jax.grad
-                ... def square(tree):
-                ...     tree = pytc.tree_unfreeze(tree)
-                ...     return tree[0]**2
-                >>> tree = (1., 2)  # contains a non-differentiable node
-                >>> square(pytc.tree_freeze(tree))
-                (Array(2., dtype=float32, weak_type=True), #2)
-    """
-
-    if isinstance(mask, Callable):
-        return jax.tree_map(
-            lambda x: unfreeze(x) if mask(x) else x,
-            tree,
-            is_leaf=is_frozen,
-        )
-
-    if jtu.tree_structure(tree, is_leaf=is_frozen) == jtu.tree_structure(mask):
-        return jax.tree_map(
-            lambda x, y: unfreeze(x) if y else x,
-            tree,
-            mask,
-            is_leaf=is_frozen,
-        )
-
-    raise ValueError(
-        f"`mask` must be a callable that accepts a leaf a returns a boolean "
-        f"or a tree with the same structure as tree with boolean values."
-        f" Got {mask=} and {tree=}."
-    )
