+++ pytreeclass/_src/tree_mask.py
@@ -1,0 +1,322 @@
+# Copyright 2023 PyTreeClass authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Utilities to work with non-jax type tree leaves across JAX transformations."""
+
+from __future__ import annotations
+
+from typing import Any, Callable, Generic, NamedTuple, TypeVar
+
+import jax
+import jax.tree_util as jtu
+import numpy as np
+
+from pytreeclass._src.tree_pprint import tree_repr, tree_str
+from pytreeclass._src.tree_util import IsLeafType, is_tree_equal, tree_copy, tree_hash
+
+T = TypeVar("T")
+
+
+class _FrozneError(NamedTuple):
+    opname: str
+
+    def __call__(self, *a, **k):
+        raise NotImplementedError(
+            f"Cannot apply `{self.opname}` operation to a frozen object "
+            f"{', '.join(map(str, a))} "
+            f"{', '.join(k + '=' + str(v) for k, v in k.items())}.\n"
+            "Unfreeze the object first by unmasking the Frozen mask:\n"
+            "Example:\n"
+            ">>> import jax\n"
+            ">>> import pytreeclass as pytc\n"
+            ">>> tree = pytc.tree_unmask(tree)"
+        )
+
+
+class _Frozen(Generic[T]):
+    __slots__ = ["__wrapped__", "__weakref__"]
+    __wrapped__: T
+
+    def __init__(self, tree: T) -> None:
+        object.__setattr__(self, "__wrapped__", tree)
+
+    def __setattr__(self, _, __) -> None:
+        raise AttributeError("Cannot assign to frozen instance.")
+
+    def __delattr__(self, _: str) -> None:
+        raise AttributeError("Cannot delete from frozen instance.")
+
+    def __repr__(self) -> str:
+        return "#" + tree_repr(self.__wrapped__)
+
+    def __str__(self) -> str:
+        return "#" + tree_str(self.__wrapped__)
+
+    def __copy__(self) -> _Frozen[T]:
+        return _Frozen(tree_copy(self.__wrapped__))
+
+    def __eq__(self, rhs: Any) -> bool | jax.Array:
+        if not isinstance(rhs, _Frozen):
+            return False
+        return is_tree_equal(self.__wrapped__, rhs.__wrapped__)
+
+    def __hash__(self) -> int:
+        return tree_hash(self.__wrapped__)
+
+    # raise helpful error message when trying to interact with frozen object
+    __add__ = __radd__ = __iadd__ = _FrozneError("+")
+    __sub__ = __rsub__ = __isub__ = _FrozneError("-")
+    __mul__ = __rmul__ = __imul__ = _FrozneError("*")
+    __matmul__ = __rmatmul__ = __imatmul__ = _FrozneError("@")
+    __truediv__ = __rtruediv__ = __itruediv__ = _FrozneError("/")
+    __floordiv__ = __rfloordiv__ = __ifloordiv__ = _FrozneError("//")
+    __mod__ = __rmod__ = __imod__ = _FrozneError("%")
+    __pow__ = __rpow__ = __ipow__ = _FrozneError("**")
+    __lshift__ = __rlshift__ = __ilshift__ = _FrozneError("<<")
+    __rshift__ = __rrshift__ = __irshift__ = _FrozneError(">>")
+    __and__ = __rand__ = __iand__ = _FrozneError("and")
+    __xor__ = __rxor__ = __ixor__ = _FrozneError("")
+    __or__ = __ror__ = __ior__ = _FrozneError("or")
+    __neg__ = __pos__ = __abs__ = __invert__ = _FrozneError("unary operation")
+    __call__ = _FrozneError("__call__")
+
+
+jtu.register_pytree_node(
+    nodetype=_Frozen,
+    flatten_func=lambda tree: ((), tree),
+    unflatten_func=lambda treedef, _: treedef,
+)
+
+
+def freeze(wrapped: _Frozen[T] | T) -> _Frozen[T]:
+    """Freeze a value to avoid updating it by `jax` transformations.
+
+    Example:
+        >>> import jax
+        >>> import pytreeclass as pytc
+        >>> import jax.tree_util as jtu
+        >>> # Usage with `jax.tree_util.tree_leaves`
+        >>> # no leaves for a wrapped value
+        >>> jtu.tree_leaves(pytc.freeze(2.))
+        []
+
+        >>> # retrieve the frozen wrapper value using `is_leaf=pytc.is_frozen`
+        >>> jtu.tree_leaves(pytc.freeze(2.), is_leaf=pytc.is_frozen)
+        [#2.0]
+
+        >>> # Usage with `jax.tree_util.tree_map`
+        >>> a= [1,2,3]
+        >>> a[1] = pytc.freeze(a[1])
+        >>> jtu.tree_map(lambda x:x+100, a)
+        [101, #2, 103]
+
+    Note:
+        - `freeze` is idempotent, i.e. `freeze(freeze(x)) == freeze(x)`
+        - Use `tree_freeze` for the `tree_map` version of `freeze`.
+            Example:
+                >>> import pytreeclass as pytc
+                >>> tree = [1,2,3]
+                >>> # freezing the pytree items
+                >>> pytc.tree_freeze(tree)
+                [#1, #2, #3]
+                >>> # freeze the pytree itself
+                >>> pytc.freeze(tree)
+                #[1, 2, 3]
+    """
+    return wrapped if is_frozen(wrapped) else _Frozen(wrapped)  # type: ignore
+
+
+def is_frozen(wrapped: Any) -> bool:
+    """Returns True if the value is a frozen wrapper."""
+    return isinstance(wrapped, _Frozen)
+
+
+def unfreeze(wrapped: _Frozen[T] | T) -> T:
+    """Unfreeze `frozen` value, otherwise return the value itself.
+
+    - use `is_leaf=pytc.is_frozen` with `jax.tree_map` to unfreeze a tree.**
+
+    Example:
+        >>> import pytreeclass as pytc
+        >>> import jax
+        >>> frozen_value = pytc.freeze(1)
+        >>> pytc.unfreeze(frozen_value)
+        1
+        >>> # usage with `jax.tree_map`
+        >>> frozen_tree = jax.tree_map(pytc.freeze, {"a": 1, "b": 2})
+        >>> unfrozen_tree = jax.tree_map(pytc.unfreeze, frozen_tree, is_leaf=pytc.is_frozen)
+        >>> unfrozen_tree
+        {'a': 1, 'b': 2}
+    """
+    return getattr(wrapped, "__wrapped__") if is_frozen(wrapped) else wrapped
+
+
+def is_nondiff(wrapped: Any) -> bool:
+    """
+    Returns True if the node is a non-differentiable node, and False for if the
+    node is of type float, complex number, or a numpy array of floats or
+    complex numbers.
+
+    Example:
+        >>> import pytreeclass as pytc
+        >>> import jax.numpy as jnp
+        >>> pytc.is_nondiff(jnp.array(1))  # int array is non-diff type
+        True
+        >>> pytc.is_nondiff(jnp.array(1.))  # float array is diff type
+        False
+        >>> pytc.is_nondiff(1)  # int is non-diff type
+        True
+        >>> pytc.is_nondiff(1.)  # float is diff type
+        False
+
+    Note:
+        This function is meant to be used with `jax.tree_map` to
+        create a mask for non-differentiable nodes in a tree, that can be used
+        to freeze the non-differentiable nodes before passing the tree to a
+        `jax` transformation.
+    """
+    if hasattr(wrapped, "dtype") and np.issubdtype(wrapped.dtype, np.inexact):
+        return False
+    if isinstance(wrapped, (float, complex)):
+        return False
+    return True
+
+
+def _tree_mask_map(
+    tree: T,
+    mask: T | Callable[[Any], bool],
+    func: type | Callable[[Any], Any],
+    *,
+    is_leaf: IsLeafType = None,
+):
+    # apply func to leaves satisfying mask pytree/condtion
+    lhsdef = jtu.tree_structure(tree, is_leaf=is_leaf)
+    rhsdef = jtu.tree_structure(mask, is_leaf)
+
+    if lhsdef == rhsdef:
+        return jax.tree_map(
+            lambda x, y: func(x) if y else x,
+            tree,
+            mask,
+            is_leaf=is_leaf,
+        )
+
+    if isinstance(mask, Callable):
+        return jax.tree_map(
+            lambda x: func(x) if mask(x) else x,
+            tree,
+            is_leaf=is_leaf,
+        )
+
+    raise ValueError(
+        f"`mask` must be a callable that accepts a leaf a returns a boolean "
+        f"or a tree with the same structure as tree with boolean values."
+        f" Got {mask=} and {tree=}."
+    )
+
+
+def tree_mask(
+    tree: T,
+    mask: T | Callable[[Any], bool] = is_nondiff,
+    *,
+    wrap: type | Callable[[Any], Any] = freeze,
+    is_leaf: IsLeafType = None,
+):
+    """Mask a tree leaves by `wrap` according to `mask`
+
+    Args:
+        tree: A pytree of values.
+        mask: A pytree of boolean values or a callable that accepts a leaf and
+            returns a boolean. If a leaf is True either in the mask or the
+            callable, the leaf is wrapped by `wrapper`, otherwise it is unchanged.
+            defaults to `is_nondiff` which returns true for non-differentiable nodes.
+        wrap: A class or a callable that accepts a leaf and returns a wrapped
+            leaf. defaults to `freeze` which wraps with a wrapper that yields no
+            leaves when `jax.tree_util.tree_flatten` is called on it.
+        is_leaf: A callable that accepts a leaf and returns a boolean. If
+            provided, it is used to determine if a value is a leaf. for example,
+            `is_leaf=lambda x: isinstance(x, list)` will treat lists as leaves
+            and will not recurse into them.
+
+    Example:
+        >>> import pytreeclass as pytc
+        >>> tree = [1, 2, {"a": 3, "b": 4.}]
+        >>> # freeze all non-differentiable nodes by default
+        >>> frozen_tree = pytc.tree_mask(tree)
+        >>> frozen_tree
+        [#1, #2, {'a': #3, 'b': 4.0}]
+        >>> pytc.tree_unmask(frozen_tree)
+        [1, 2, {'a': 3, 'b': 4.0}]
+
+    Example:
+        >>> # pass non-differentiable values to `jax.grad`
+        >>> import pytreeclass as pytc
+        >>> import jax
+        >>> @jax.grad
+        ... def square(tree):
+        ...     tree = pytc.tree_unmask(tree)
+        ...     return tree[0]**2
+        >>> tree = (1., 2)  # contains a non-differentiable node
+        >>> square(pytc.tree_mask(tree))
+        (Array(2., dtype=float32, weak_type=True), #2)
+    """
+    return _tree_mask_map(tree, mask=mask, func=wrap, is_leaf=is_leaf)
+
+
+def tree_unmask(
+    tree: T,
+    mask: T | Callable[[Any], bool] = lambda _: True,
+    *,
+    unwrap: Callable[[Any], Any] = unfreeze,
+    is_wrap: Callable[[Any], bool] = is_frozen,
+):
+    """Unmask a tree leaves by `unwrap` according to `mask`
+
+    Args:
+        tree: A pytree of values.
+        mask: A pytree of boolean values or a callable that accepts a leaf and
+            returns a boolean. If a leaf is True either in the mask or the
+            callable, the leaf is unwrapped by `unwrap`, otherwise it is unchanged.
+            defaults to `lambda _: True` which returns true for all nodes.
+        unwrap: A callable that accepts a leaf and returns an unwrapped leaf.
+            defaults to `unfreeze` which unwraps a frozen leaf.
+        is_wrap: A callable that accepts a leaf and returns a boolean. to
+            determine if a value is wrapped by `wrap`. defaults to `is_frozen`
+            which returns true for frozen nodes.
+
+    Example:
+        >>> import pytreeclass as pytc
+        >>> tree = [1, 2, {"a": 3, "b": 4.}]
+        >>> # freeze all non-differentiable nodes by default
+        >>> frozen_tree = pytc.tree_mask(tree)
+        >>> frozen_tree
+        [#1, #2, {'a': #3, 'b': 4.0}]
+        >>> pytc.tree_unmask(frozen_tree)
+        [1, 2, {'a': 3, 'b': 4.0}]
+
+    Example:
+        >>> # pass non-differentiable values to `jax.grad`
+        >>> import pytreeclass as pytc
+        >>> import jax
+        >>> @jax.grad
+        ... def square(tree):
+        ...     tree = pytc.tree_unmask(tree)
+        ...     return tree[0]**2
+        >>> tree = (1., 2)  # contains a non-differentiable node
+        >>> square(pytc.tree_mask(tree))
+        (Array(2., dtype=float32, weak_type=True), #2)
+    """
+    # on backend this is exactly the same as `tree_mask`, but with different defaults
+    # and `is_wrap` is named of `is_leaf to make it easier for the end-user.
+    return _tree_mask_map(tree, mask=mask, func=unwrap, is_leaf=is_wrap)
