+++ pytreeclass/_src/tree_pprint.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""Utilities for pretty printing pytrees."""
+
 from __future__ import annotations
 
 import dataclasses as dc
